Refactor: Moved addVideo components into components folder
